% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqs_user.R
\name{aqs_signup}
\alias{aqs_signup}
\title{Create an account for the AQS API}
\usage{
aqs_signup(email)
}
\arguments{
\item{email}{A string specifying an email account to register as a user}
}
\value{
No return value, called to sign up for the AQS API
}
\description{
This function helps you create an account or reset a password. Once you
execute this function, a verification email will be sent to the email
account specified. If the request is made with an email that is already
registered, a new key will be issued for that account and emailed to the
listed address.
}
\examples{
\dontrun{

## Please use your email address to create an account

aqs_signup(email = "youremail@toregister.com")
}

}
\seealso{
See \link{set_aqs_user} to set your credentials to send a request to the
AQS API.
}
