% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list}
\alias{list}
\alias{list_states}
\alias{list_counties_by_state}
\alias{list_sites_by_county}
\alias{list_cbsas}
\alias{list_classes}
\alias{list_parameters_by_class}
\alias{list_pqaos}
\alias{list_mas}
\title{List}
\usage{
list_states()

list_counties_by_state(state)

list_sites_by_county(state, county)

list_cbsas()

list_classes()

list_parameters_by_class(class)

list_pqaos()

list_mas()
}
\arguments{
\item{state}{The 2 digit state FIPS code for the state (with leading zero). They may be obtained via the list states service. Only data from within this state will be returned.}

\item{county}{The 3 digit state FIPS code for the county within the state (with leading zeroes). They may be obtained via the list counties service. Only data from within this county will be returned.}

\item{class}{The AQS parameter class name. Classes are a way to group like parameters. They may be obtained via the list classes service. Only used in the list parameters service.}
}
\description{
Provides information you need to construct other queries. Valid values for the required variables: parameter code, state code, etc. (See below.)
}
