% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{meta_data}
\alias{meta_data}
\alias{meta_data_is_available}
\alias{meta_data_fields_by_service}
\alias{meta_data_issues}
\title{Meta Data}
\usage{
meta_data_is_available()

meta_data_fields_by_service(service)

meta_data_fields_by_service(service)

meta_data_issues()
}
\description{
Returns information about the API. Is it available/up? What are the meanings of the fields in my returned data? Recent and planned changes, etc.
}
