% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitors.R
\name{monitors}
\alias{monitors}
\alias{monitors_by_site}
\alias{monitors_by_county}
\alias{monitors_by_state}
\alias{monitors_by_box}
\alias{monitors_by_cbsa}
\title{Monitors}
\usage{
monitors_by_site(param, bdate, edate, state, county, site)

monitors_by_county(param, bdate, edate, state, county)

monitors_by_state(param, bdate, edate, state)

monitors_by_box(param, bdate, edate, minlat, maxlat, minlon, maxlon)

monitors_by_cbsa(param, bdate, edate, cbsa)
}
\arguments{
\item{param}{The AQS parameter code for the data selection. AQS uses proprietary 5 digit codes. They may be obtained via the list parameters service. Up to 5 parameters may be requested, separated by commas. Only data for these parameters will be returned.}

\item{bdate}{The begin date of the data selection in YYYYMMDD format. Only data on or after this date will be returned. (Note, for annual data, only the year portion of the bdate and edate are used and only whole years of data are returned. For example, bdate = 20171231 and edate = 20180101 will return full data for 2017 and 2018.)}

\item{edate}{The end date of the data selection in YYYYMMDD format. Only data on or before this date will be returned. The end date must be in the same year as the begin date. This rule does not apply to the Monitors service. (Note, for annual data, only the year portion of the bdate and edate are used and only whole years of data are returned. For example, bdate = 20170531 and edate = 20170630 will return full data for 2017.)}

\item{state}{The 2 digit state FIPS code for the state (with leading zero). They may be obtained via the list states service. Only data from within this state will be returned.}

\item{county}{The 3 digit state FIPS code for the county within the state (with leading zeroes). They may be obtained via the list counties service. Only data from within this county will be returned.}

\item{site}{The 4 digit AQS site number within the county (with leading zeroes). They may be obtained via the list sites service. Only data from this site will be returned.}

\item{minlat}{The minimum latitude of a geographic box. Decimal latitude with north being positive. Only data north of this latitude will be returned.}

\item{maxlat}{The maximum latitude of a geographic box. Decimal latitude with north being positive. Only data south of this latitude will be returned.}

\item{minlon}{The minimum longitude of a geographic box. Decimal longitude with east being positive. Only data east of this longitude will be returned.}

\item{maxlon}{The maximum longitude of a geographic box. Decimal longitude with east being positive. Only data west of this longitude will be returned. Note that -80 is less than -70.}

\item{cbsa}{The 5 digit AQS Core Based Statistical Area code (as defined by the census bureau, with leading zeroes). They may be obtained via the list cbsas service. Only data from this area will be returned.}
}
\description{
Returns operational information about the samplers (monitors) used to collect the data. Includes identifying information, operational dates, operating organizations, etc.
}
