% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_flow_rate_verifications.R
\name{qa_flow_rate_verifications}
\alias{qa_flow_rate_verifications}
\alias{qa_flow_rate_verifications_by_site}
\alias{qa_flow_rate_verifications_by_county}
\alias{qa_flow_rate_verifications_by_state}
\alias{qa_flow_rate_verifications_by_pqao}
\alias{qa_flow_rate_verifications_by_ma}
\title{QA Flow Rate Verifications}
\usage{
qa_flow_rate_verifications_by_site(param, bdate, edate, state, county, site)

qa_flow_rate_verifications_by_county(param, bdate, edate, state, county)

qa_flow_rate_verifications_by_state(param, bdate, edate, state)

qa_flow_rate_verifications_by_pqao(param, bdate, edate, pqao)

qa_flow_rate_verifications_by_ma(param, bdate, edate, ma)
}
\arguments{
\item{param}{The AQS parameter code for the data selection. AQS uses proprietary 5 digit codes. They may be obtained via the list parameters service. Up to 5 parameters may be requested, separated by commas. Only data for these parameters will be returned.}

\item{bdate}{The begin date of the data selection in YYYYMMDD format. Only data on or after this date will be returned. (Note, for annual data, only the year portion of the bdate and edate are used and only whole years of data are returned. For example, bdate = 20171231 and edate = 20180101 will return full data for 2017 and 2018.)}

\item{edate}{The end date of the data selection in YYYYMMDD format. Only data on or before this date will be returned. The end date must be in the same year as the begin date. This rule does not apply to the Monitors service. (Note, for annual data, only the year portion of the bdate and edate are used and only whole years of data are returned. For example, bdate = 20170531 and edate = 20170630 will return full data for 2017.)}

\item{state}{The 2 digit state FIPS code for the state (with leading zero). They may be obtained via the list states service. Only data from within this state will be returned.}

\item{county}{The 3 digit state FIPS code for the county within the state (with leading zeroes). They may be obtained via the list counties service. Only data from within this county will be returned.}

\item{site}{The 4 digit AQS site number within the county (with leading zeroes). They may be obtained via the list sites service. Only data from this site will be returned.}

\item{pqao}{The 4 digit AQS Primary Quality Assurance Organization code (with leading zeroes). They may be obtained via the list pqaos service. Only data for this PQAO will be returned. Only used in QA Data services.}

\item{ma}{The 4 digit AQS Monitoring Agency code (with leading zeroes). They may be obtained via the list mas service. Only data for this Monitoring Agency will be returned. Only used in QA Data services.}
}
\description{
Quality assurance data - flow rate verifications. Several times per year, each PM monitor must have it's (fixed) flow rate verified by an operator taking a measurement of the flow rate.
}
