% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accountManagement.R
\name{setAQSUser}
\alias{setAQSUser}
\title{setAQSUser}
\usage{
setAQSUser(email, key)
}
\arguments{
\item{email}{The email you used to sign up for AQS}

\item{key}{The access key provided by AQS}
}
\description{
Function to store your AQS credentials so they can be used by raqs
}
