% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_up.R
\name{sign_up}
\alias{sign_up}
\alias{signup}
\title{Sign up}
\usage{
signup(email)
}
\arguments{
\item{email}{The email address of the requester.}
}
\description{
Creates an account or resets a password. Requires validation of registered email address. Email will be sent to the registered address from aqsdatamart@epa.gov. This service can also be used to reset passwords.
}
